refactor: 서비스 인터페이스 삭제
- 구현체 이름을 CalculatorService로 변경
